+++ src/Controller/Common/DeleteUserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controller\Common;
 
+use App\Entity\Preferences;
 use App\Entity\Token;
 use App\Entity\User;
 use Doctrine\Persistence\ManagerRegistry;
@@ -11,10 +12,20 @@ class DeleteUserController extends AbstractApiController
 
     public function invoke(User $user, ManagerRegistry $doctrine)
     {
+        $entityManager = $doctrine->getManager();
+
         // remove tokens
 
         $rep = $doctrine->getRepository(Token::class);
         $rep->removeByUser($user);
+
+        // remove preferences
+
+        $preferencesRep = $doctrine->getRepository(Preferences::class);
+        $preferences = $preferencesRep->findByUser($user);
+        if ($preferences) {
+            $entityManager->remove($preferences);
+        }
 
         // remove all user informations
 
@@ -24,7 +35,6 @@ class DeleteUserController extends AbstractApiController
             ->setRoles([])
             ->setDeleted(true);
 
-        $entityManager = $doctrine->getManager();
         $entityManager->persist($user);
         $entityManager->flush();
     }